requirements.txt
matplotlib==3.8.0
requests==2.31.0
python-dotenv==1.0.0 